\name{gm}
\alias{gm}

\title{
Generic Generator Matrix Estimation
}
\description{
Generic function for generator matrix estimation
}
\usage{
gm(tm, ...)
}

\arguments{
  \item{tm}{
matrix of either absolute or relative transition frequencies
}
  \item{\dots}{
additional arguments
}
}

\seealso{
\code{\link{gm.default}}
}
