\name{gmci}
\alias{gmci}
\title{
Generic Generator Matrix Confidence / Credibility Interval
}
\description{
Generic function for generator matrix confidence / credibility intervals
}
\usage{
gmci(gm, ...)
}
\arguments{
  \item{gm}{
a generator matrix estimation object
}
  \item{\dots}{
additional arguments
}
}\seealso{
\code{\link{gmci.default}}
}