\name{print.gm}
\alias{print.gm}

\title{
Print Method for Generator Matrix Estimation Objects
}
\description{
Function for printing the results of a generator matrix estimation}
\usage{
\method{print}{gm}(x, ...)
}

\arguments{
  \item{x}{
a generator matrix estimation object
}
  \item{\dots}{
additional arguments
}
}

\value{
generator matrix
}

\seealso{
\code{\link{summary.gm}}, \code{\link{plot.gm}}
}