\name{print.gmci}
\alias{print.gmci}
\title{
Print Method for Generator Matrix Confidence / Credibility Interval Objects
}
\description{
Function for printing the boundaries of a generator matrix confidence / credibility interval}
\usage{
\method{print}{gmci}(x, ...)
}
\arguments{
  \item{x}{
a generator matrix confidence / credibility interval
}
  \item{\dots}{
additional arguments
}
}

\value{
generator matrix confidence bounds
}

\seealso{
\code{\link{plot.gmci}}
}