\name{summary.gm}
\alias{summary.gm}

\title{
Extended Output for Generator Matrix Estimate Objects
}
\description{
Function for providing results and extended output of a generator matrix estimation procedure.
}
\usage{
\method{summary}{gm}(object, ...)
}

\arguments{
  \item{object}{
a generator matrix estimation object
}
  \item{\dots}{
additional arguments
}
}

\value{
estimation summary
}

\seealso{
\code{\link{print.gm}}, \code{\link{plot.gm}}
}